Move checkCustomerEmail UI behaviour back to the app
