+++ web/app/containers/checkout-shipping/reducer.js
@@ -1,11 +1,12 @@
 import {handleActions} from 'redux-actions'
 import Immutable from 'immutable'
 import {mergePayloadForActions} from '../../utils/reducer-utils'
-import {showCompanyAndApt} from './actions'
-import {receiveCheckoutShippingData} from '../../integration-manager/responses'
+import {showCompanyAndApt, receiveCheckoutShippingData} from './actions'
+import * as integrationManagerResponses from '../../integration-manager/checkout/responses'
+
 
 export default handleActions({
-    ...mergePayloadForActions(receiveCheckoutShippingData),
+    ...mergePayloadForActions(integrationManagerResponses.receiveCheckoutShippingData, receiveCheckoutShippingData),
     [showCompanyAndApt]: (state) => {
         return state.merge({isCompanyOrAptShown: true})
     }
